refactor(backend): layered routes, controllers, models, config
Split monolithic api.js into controllers; move data access to models/repositories; Google OAuth in config; agent prompt helpers in services; mountExpressApp + server.js entry. JSON API uses Routes→Controllers→Services/Models (no classic View). package.json start/dev use server/server.js; index.js re-exports for compatibility.
